advanced/project/forecast_bot/requirements.txt
aiogram==3.22.0
aiohttp
aiosqlite
python-dotenv
pytest
pytest-asyncio
httpx
